rename prioritizedRoute() to featureRoute() -and- improve priority discussion in docs
