+++ src/feature/discovery/route.js
@@ -1,6 +1,6 @@
 import React                  from 'react';
 import * as sel               from './state';
-import {prioritizedRoute, 
+import {featureRoute, 
         PRIORITY}             from '../../util/feature-u/aspect/feature-u-state-router';
 import featureName            from './featureName';
 import DiscoveryFilterScreen  from './comp/DiscoveryFilterScreen';
@@ -13,7 +13,7 @@ import DiscoveryListScreen    from './comp/DiscoveryListScreen';
 
 export default [
 
-  prioritizedRoute({
+  featureRoute({
     priority: PRIORITY.HIGH,
     content({app, appState}) {
       // display DiscoveryFilterScreen, when form is active (accomplished by our logic)
@@ -25,7 +25,7 @@ export default [
     }
   }),
 
-  prioritizedRoute({
+  featureRoute({
   //priority: PRIORITY.STANDARD,
     content({app, appState}) {
